+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.12.0 - 2019-07-21
+### Added
+- Tests of next event after calling `getValue()`, `getArray()`, and `getObject()`.
+- Tests of reading JSON followed by garbages.
+- Add tests for `read()`, `readArray()`, and `readObject()` of `JsonReader`.
+
 ## 0.11.0 - 2019-07-20
 ### Added
 - More tests for `JsonParser.hasNext()`.